Parse stratified survival data inside SurvivalPlot
