+++ src/GuppyDataExplorer/ExplorerSurvivalAnalysis/index.jsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import SurvivalPlot from './SurvivalPlot';
 import ControlForm from './ControlForm';
 import RiskTable from './RiskTable';
-import { getSurvivalSeries } from './utils';
 import {
   factors as mockFactors,
   fetchResult as fetchMockSurvivalResult,
@@ -12,16 +11,18 @@ import './ExplorerSurvivalAnalysis.css';
 function ExplorerSurvivalAnalysis({ aggsData, filters }) {
   const [pval, setPval] = useState();
   const [risktable, setRisktable] = useState([]);
-  const [survivalSeries, setSurvivalSeries] = useState([]);
+  const [survival, setSurvival] = useState([]);
+  const [stratificationVariable, setStratificationVariable] = useState('');
   const [timeInterval, setTimeInterval] = useState(2);
 
   const handleSubmit = (userInput) => {
+    setStratificationVariable(userInput.stratificationVariable);
     setTimeInterval(userInput.timeInterval);
 
     fetchMockSurvivalResult(userInput).then((result) => {
       setPval(result.pval && +parseFloat(result.pval).toFixed(4));
       setRisktable(result.risktable);
-      setSurvivalSeries(getSurvivalSeries(result.survival, userInput));
+      setSurvival(result.survival);
     });
   };
 
@@ -42,7 +43,11 @@ function ExplorerSurvivalAnalysis({ aggsData, filters }) {
         <div className='explorer-survival-analysis__pval'>
           {pval && `Log-rank test p-value: ${pval}`}
         </div>
-        <SurvivalPlot data={survivalSeries} timeInterval={timeInterval} />
+        <SurvivalPlot
+          data={survival}
+          stratificationVariable={stratificationVariable}
+          timeInterval={timeInterval}
+        />
         <RiskTable data={risktable} timeInterval={timeInterval} />
       </div>
     </div>
