+++ src/GuppyDataExplorer/ExplorerSurvivalAnalysis/utils.js
@@ -1,15 +1,3 @@
-export const getSurvivalSeries = (survival, userInput) =>
-  userInput.factorVariable === '' || userInput.stratificationVariable === ''
-    ? survival
-    : survival.reduce((acc, { name, data }) => {
-        const [factorKey, stratificationKey] = name.split(',');
-        const stratificationValue = acc.hasOwnProperty(stratificationKey)
-          ? [...acc[stratificationKey], { name: factorKey, data }]
-          : [{ name: factorKey, data }];
-
-        return { ...acc, [stratificationKey]: stratificationValue };
-      }, {});
-
 export const getXAxisTicks = (data, step = 2) => {
   const times = data.flatMap(({ data }) => data).map(({ time }) => time);
   const minTime = Math.floor(Math.min(...times));
